+++ app/controllers/application_controller.rb
@@ -9,6 +9,10 @@ class ApplicationController < ActionController::Base
     redirect_to root_path, alert: "You do not have sufficient privileges to access that page." unless current_user.system_admin?
   end
 
+  def team
+    @team_members = Admin::TeamMember.order('position')
+  end
+
   protected
 
     def configure_permitted_parameters